+++ lib/constance.h
@@ -1,0 +1,12 @@
+#ifndef _CONSTANCE_H
+#define _CONSTANCE_H
+
+#define DEFAULT_PORT 80
+#define MAX_RESPONSE_SIZE (1024 * 1024 * 5)
+#define MAX_REQUEST_SIZE (1024)
+#define MAX_HOSTNAME_SIZE 1024
+#define MAX_PATH_SIZE 1024
+#define MAX_SIZE 128
+#define MAX_URL 256
+
+#endif
